Registration page
The login and registration buttons are both present on the login page. If the user presses the login button, they get taken to the login page. If the user presses registration they get taken to the registration page. But before pressing on the login or registration page, they need to first select if they are logging in/ registering as a donor or receiver.